backend/requirements.txt
Django==5.0.8
djangorestframework==3.14.0
django-filter==23.3
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
django-cors-headers==4.3.1
Pillow==10.4.0
requests==2.32.3
geoip2==4.8.0
redis==5.0.1
gunicorn==22.0.0
whitenoise==6.7.0
stripe==11.4.1
resend==2.4.0